rxparamsetup: Move EU868 rxparamsetup tests to own file
